chore: reorganized files and directories
